The main purpose of the revision is to improve the performance of ordered display groups.  For instance, Q3Group_AddObject will now take constant time, rather than time linear in the number of items in the group.
I fixed E3DisplayGroup_GetType so that if you give it an object of
type kQ3GroupTypeDisplay, it will return kQ3GroupTypeDisplay instead
of kQ3ObjectTypeInvalid.

In e3group_createPosition and e3group_setposition, if the object is
not acceptable, post an error.

In e3group_getnextobjectposition and e3group_getprevobjectposition, it said
	TQ3XGroupPosition		*pos = (TQ3XGroupPosition*)position;
but should have said
	TQ3XGroupPosition		*pos = (TQ3XGroupPosition*)*position;

Display groups now accept only drawable objects, as documented.
--
   James W. Walker, ScriptPerfection Enterprises, Inc.