Fix unused inputs in GLSL fragment shaders causing crash on Windows/NVIDIA [from Jorio fork]
When using NULLIllumination or flat shading, the generated shader does not use normals. Quesa would still try to pass normals to the shader, which would result in a GL_INVALID_VALUE error.

There was another issue where using Lambert or NULL illumination on non-texturemapped meshes would crash as well, due to Quesa passing texCoords to a shader not using it. This issue seems to have been resolved by this fix.