More Swift-friendly defines (#13)
Objective-C  and Swift can also have enums be an explicit size. Checking for objc_fixed_enum feature should be good enough to expose the enums.
Swift hates it if inttypes.h is included: include stdint.h if we are using a compiler that support c99 or later or C++11 or later.